refactor(keymap): split KeyMap → ttymap-tui, KeybindingOverrides → ttymap-config
Each type lands where it semantically belongs: the live KeyMap
beside the rest of the input subsystem (where the crossterm dep
that backs KeyCode already lives), and KeybindingOverrides
beside the rest of the user-facing settings shape. ttymap-core
becomes pure cross-cutting vocabulary (UserCommand + EventBus) —
crossterm dep dropped.

ttymap-tui picks up a ttymap-config edge; the "tui must be
config-free" self-imposed constraint had no concrete consumer
driving it, so the linear DAG (engine → core → config → tui →
lua → cli → app) is preferable to the previous parallel-branches
shape.

UserCommand::keymap_help_entries(&KeyMap) becomes the inherent
KeyMap::help_entries() method.
