+++ ttymap-app/src/app/mod.rs
@@ -48,9 +48,9 @@ use log::{debug, info};
 use self::overlay::OverlayThrottle;
 use self::sidebar::SidebarPolicy;
 use ttymap_config::Config;
+pub use ttymap_config::KeybindingOverrides;
 use ttymap_core::UserCommand;
 use ttymap_core::event::{Event, EventBus};
-pub use ttymap_core::keymap::KeybindingOverrides;
 use ttymap_engine::map::render::frame::MapFrame;
 use ttymap_engine::map::state::MapState;
 use ttymap_lua::{LuaHandle, LuaSubsystem};
