+++ ttymap-tui/Cargo.toml
@@ -10,6 +10,7 @@ repository.workspace = true
 [dependencies]
 ttymap-engine = { path = "../ttymap-engine", version = "0.1.0" }
 ttymap-core = { path = "../ttymap-core", version = "0.1.0" }
+ttymap-config = { path = "../ttymap-config", version = "0.1.0" }
 crossterm = "0.29"
 ratatui = "0.30"
 unicode-width = "0.2"
